remove contracts
