+++ typings/contracts.ts
@@ -1,8 +1,5 @@
 import { ContractRegistryContract } from "./contract-registry-contract";
 import { ERC20Contract } from "./erc20-contract";
-import { StakingRewardsContract } from "./staking-rewards-contract";
-import { BootstrapRewardsContract } from "./bootstrap-rewards-contract";
-import { FeesContract } from "./fees-contract";
 import { ElectionsContract } from "./elections-contract";
 import { SubscriptionsContract } from "./subscriptions-contract";
 import { ProtocolContract } from "./protocol-contract";
@@ -22,9 +19,6 @@ import {RewardsContract} from "./rewards-contract";
 export type Contracts = {
     ContractRegistry: ContractRegistryContract & Contract;
     TestingERC20: ERC20Contract & Contract;
-    StakingRewards: StakingRewardsContract & Contract;
-    BootstrapRewards: BootstrapRewardsContract & Contract;
-    Fees: FeesContract & Contract;
     Rewards: RewardsContract & Contract;
     Elections: ElectionsContract & Contract;
     Subscriptions: SubscriptionsContract & Contract;
