+++ typings/rewards-contract.d.ts
@@ -2,6 +2,45 @@ import {TransactionConfig, TransactionReceipt} from "web3-core";
 import {Contract} from "../eth";
 import * as BN from "bn.js";
 
+export interface BootstrapRewardsAssignedEvent {
+    assignees: string[],
+    amounts: Array<string|BN>
+}
+
+export interface BootstrapAddedToPoolEvent {
+    added: number|BN,
+    total: number|BN
+}
+
+export interface FeesAddedToBucketEvent {
+    bucketId: string|BN,
+    added: string|BN,
+    total: string|BN,
+    isCompliant: boolean
+}
+
+export interface FeesAssignedEvent {
+    assignees: string[],
+    orbs_amounts: Array<string|BN>
+}
+
+export interface StakingRewardAssignedEvent {
+    assignee: string,
+    amount: string|BN,
+    balance: string|BN
+}
+
+export interface StakingRewardsDistributedEvent {
+    distributer: string,
+    fromBlock: string|BN,
+    toBlock: string|BN,
+    split: string|BN,
+    txIndex: string|BN,
+    to: string[],
+    amounts: (string|BN)[]
+}
+
+
 export interface RewardsContract extends Contract {
     // staking rewards
     distributeOrbsTokenStakingRewards(totalAmount: (number|BN), fromBlock: (number|BN), toBlock: (number|BN), split: (number|BN), txIndex: (number|BN), to: string[], amounts: (number | BN)[], params?: TransactionConfig): Promise<TransactionReceipt>;
